workaround for server sync issues
server sends confirmation notification before the data really is
changed. As a result when one does quick follow-up query for the data
that just got created the result is unexpected.
As a workaround this commit adds sleeps to several parts of the code
which typically cause test failures, making tests more reliable